% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{push_data}
\alias{push_data}
\title{Send data to a Flight server}
\usage{
push_data(client, data, path)
}
\arguments{
\item{client}{`pyarrow.flight.FlightClient`, as returned by [flight_connect()]}

\item{data}{`data.frame` or [arrow::RecordBatch] to upload}

\item{path}{string identifier to store the data under}
}
\value{
`client`, invisibly.
}
\description{
Send data to a Flight server
}
